requirements.txt
torch>=2.0.0
torchvision
ultralytics
segment-anything
transformers
faiss-cpu
opencv-python
Pillow
streamlit
PyYAML
tqdm
